Move _switching up out of core.
This silences a warning that were importing an private
module from fixate.core in the top level __init__.py. So
partly this is just to silence PyCharm, but it is a resonable
change in general.
